refactoring grobid client generic
